Robust Blind Verification

Package documentation

Docs for package

RobustBlindVerification.add_noise! — Method

add_noise!(client::Client, noise_model::NoiseModel)

This function adds noise to a quantum system according to the specified noise model. It calls the add_noise! function with a new client and the given noise model as arguments.

Arguments

client::Client: The client for which the noise is being added.

noise_model::NoiseModel: The noise model to be used.

Examples

client = Client()
noise_model = NoiseModel()
add_noise!(client, noise_model)

RobustBlindVerification.add_noise! — Method

add_noise!(client::Client, model::Union{Damping,Dephasing,Depolarising,Pauli}, params::QubitNoiseParameters)

This function adds noise to a quantum system. The noise model can be one of Damping, Dephasing, Depolarising, or Pauli. The function throws an error if the noise type is not SingleQubit. It then iterates over the range of qubits represented in the backend and adds noise to each qubit according to the specified model and parameters.

Arguments

client::Client: The client for which the noise is being added.

model::Union{Damping,Dephasing,Depolarising,Pauli}: The noise model to be used.

params::QubitNoiseParameters: The parameters for the noise model.

Examples

client = Client()
model = Damping()
params = QubitNoiseParameters(SingleQubit(), backend)
add_noise!(client, model, params)

RobustBlindVerification.c_iterator — Method

c_iterator(N)

Create a C based index iterator that generates numbers from 0 to N-1.

Arguments

N: The upper limit for the circular iterator.

Returns

A circular iterator that generates numbers from 0 to N-1.

Examples

# Create a circular iterator
N = 5
iterator = c_iterator(N)

RobustBlindVerification.c_shift_index — Method

c_shift_index(n::Int)

Compute the shifted index n-1 for circular indexing.

Arguments

n::Int: The input index.

Returns

The shifted index n-1 for circular indexing.

Examples

# Compute the shifted index
n = 3
shifted_index = c_shift_index(n)

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Computation ∩ Qubit ∈ Input set → δᵥ = ϕᵥ + (θᵥ + xᵥπ) + rᵥπ

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Computation ∩ Qubit ∉ Input set → δᵥ = ϕᵥ′ + θᵥ + rᵥπ

RobustBlindVerification.compute_angle_δᵥ — Method

MBQC compute updated angle

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Test ∩ Qubit ≡ Dummy → δᵥ = {kπ/r | k ∼ U(0..7)}

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Test ∩ Qubit ≡ Trap → δᵥ = θᵥ + rᵥπ

RobustBlindVerification.create_quantum_env — Method

create_quantum_env(client::Client)

Create a new quantum environment using the QuEST environment creation function.

Arguments

client::Client: The client for which the quantum environment is being created.

Returns

A new quantum environment.

Examples

client = Client()
env = create_quantum_env(client)

RobustBlindVerification.create_quantum_state — Method

create_quantum_state(client::Client, density_matrix::DensityMatrix, quantum_env, num_qubits)

Create a quantum state using a density matrix representation.

Arguments

client::Client: The client object representing the quantum computing service.

density_matrix::DensityMatrix: The density matrix object representing the quantum state.

quantum_env: The quantum environment object.

num_qubits: The number of qubits to be used in the quantum state.

Returns

The quantum state represented as a density matrix.

Examples

client = Client()
env = create_quantum_env(client)
state = create_quantum_state(client, DensityMatrix(), env, 2)

RobustBlindVerification.create_quantum_state — Method

create_quantum_state(client::Client, state::StateVector, quantum_env, num_qubits)

Create a new quantum state as a state vector using the QuEST function createQureg.

Arguments

client::Client: The client for which the quantum state is being created.

state::StateVector: Indicates that the quantum state should be created as a state vector.

quantum_env: The quantum environment in which the quantum state is being created.

num_qubits: The number of qubits in the quantum state.

Returns

A new quantum state as a state vector.

Examples

client = Client()
env = create_quantum_env(client)
state = create_quantum_state(client, StateVector(), env, 2)

RobustBlindVerification.draw_bit — Method

Draw random bit 0 or 1

RobustBlindVerification.draw_dᵥ — Method

Draw random bit 0 or 1 for dummy

RobustBlindVerification.draw_rᵥ — Method

Draw random bit 0 or 1 for trap

RobustBlindVerification.draw_θᵥ — Method

For draw θ a multiple of kπ/4, k ∈ 0..7

RobustBlindVerification.generate_random_greedy_color — Method

Use Graphs.randomgreedycolor(g, reps) -> Graphs.Coloring{Int64} Uses a greedy approximation to get a colouring of a graph

RobustBlindVerification.get_edge_iterator — Method

get_edge_iterator(resource::MBQCResourceState)

Retrieve an iterator over the edges in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.

# Returns
An iterator over the edges of the resource state's graph.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the edge iterator
edge_iterator = get_edge_iterator(resource)
```

RobustBlindVerification.get_number_vertices — Method

get_number_qubits(resource::MBQCResourceState)

Retrieve the number of qubits in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.

# Returns
The number of qubits in the resource state.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the number of qubits
num_qubits = get_number_qubits(resource)
```

RobustBlindVerification.get_random_coloring — Method

Extracts from Vector{Vector{Int64}} -> Vector{Int64} Draw uniform one colouring arangments

RobustBlindVerification.get_vertex_iterator — Method

get_vertex_iterator(resource::MBQCResourceState)

Retrieve an iterator over the vertices in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.

# Returns
An iterator over the vertices of the resource state's graph.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the vertex iterator
vertex_iterator = get_vertex_iterator(resource)
```

RobustBlindVerification.get_vertex_neighbours — Method

get_vertex_neighbours(resource::MBQCResourceState, vertex)

Retrieve the neighbors of a given vertex in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.
- `vertex`: The vertex for which to retrieve the neighbors.

# Returns
An array of vertices representing the neighbors of the specified vertex.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the neighbors of a vertex
vertex = 1
neighbors = get_vertex_neighbours(resource, vertex)
```

RobustBlindVerification.init_plus_phase_state! — Method

init_plus_phase_state!(nophase::NoPhase, qureg, qᵢ)

This function initializes a quantum state to a superposition state (|+⟩ state) without adding any phase. It applies a Hadamard gate to the qubit, putting it into a superposition state.

Arguments

nophase::NoPhase: The NoPhase object indicating that no phase is to be added.

qureg: The quantum register containing the qubit.

qᵢ: The index of the qubit in the quantum register.

Examples

nophase = NoPhase()
qureg = createQureg(1, env)
qᵢ = 1
init_plus_phase_state!(nophase, qureg, qᵢ)

RobustBlindVerification.init_plus_phase_state! — Method

init_plus_phase_state!(phase::Phase, qureg, qᵢ, φᵢ)

This function initializes a quantum state to a superposition state (|+⟩ state) with a specified phase. It first applies a Hadamard gate to the qubit, putting it into a superposition state. Then it applies a Z rotation to the qubit, adding the specified phase.

Arguments

phase::Phase: The phase object.

qureg: The quantum register containing the qubit.

qᵢ: The index of the qubit in the quantum register.

φᵢ: The phase to be added to the qubit.

Examples

phase = Phase()
qureg = createQureg(1, env)
qᵢ = 1
φᵢ = π/2
init_plus_phase_state!(phase, qureg, qᵢ, φᵢ)

RobustBlindVerification.init_qubit — Method

For use to initialise the qubit in the meta graph
The state is not given, but the bit for the initial 
state of the dummy

RobustBlindVerification.init_qubit — Method

For use to initialise the qubit in the meta graph
The state is not given, but the angle for the plus 
phase state for a trap qubit

RobustBlindVerification.initialise_qubit — Method

initialise_qubit(dummy::DummyQubit, noinput::NoInputQubits, quantum_state, qubit_index, qubit_input_value::Int)

This function initialises a qubit in a quantum state. If the input value for the qubit is zero, it does nothing. If the input value is one, it applies a Pauli-X gate to the qubit. If the input value is neither zero nor one, it throws a DummyQubitZeroOneInitialisationError.

Arguments

dummy::DummyQubit: The DummyQubit object.

noinput::NoInputQubits: The NoInputQubits object.

quantum_state: The quantum state containing the qubit.

qubit_index: The index of the qubit in the quantum state.

qubit_input_value::Int: The input value for the qubit.

Examples

dummy = DummyQubit()
noinput = NoInputQubits()
quantum_state = createQureg(1, env)
qubit_index = 1
qubit_input_value = 0
initialise_qubit(dummy, noinput, quantum_state, qubit_index, qubit_input_value)

RobustBlindVerification.initialise_qubit — Method

initialise_qubit(mbqc::MBQC, qubit::Union{ComputationQubit,TrapQubit}, input::Union{InputQubits,InputQubits,NoInputQubits}, quantum_state, qubit_index)

This function initialises a qubit in a quantum state to a superposition state without adding any phase. It applies a Hadamard gate to the qubit, putting it into a superposition state.

Arguments

mbqc::MBQC: The MBQC object.

qubit::Union{ComputationQubit,TrapQubit}: The type of the qubit.

input::Union{InputQubits,InputQubits,NoInputQubits}: The input object.

quantum_state: The quantum state containing the qubit.

qubit_index: The index of the qubit in the quantum state.

Examples

mbqc = MBQC()
qubit = ComputationQubit()
input = InputQubits()
quantum_state = createQureg(1, env)
qubit_index = 1
initialise_qubit(mbqc, qubit, input, quantum_state, qubit_index)

RobustBlindVerification.initialise_qubit — Method

initialise_qubit(qubit::Union{ComputationQubit,TrapQubit}, input::Union{InputQubits,InputQubits,NoInputQubits}, quantum_state, qubit_index, qubit_input_value::Float64)

This function initialises a qubit in a quantum state with a phase determined by the input value. If the input value is a float, it initialises the qubit to a superposition state with the input value as the phase. If the input value is not a float, it throws a QubitFloatPhaseInitialisationError.

Arguments

qubit::Union{ComputationQubit,TrapQubit}: The type of the qubit.

input::Union{InputQubits,InputQubits,NoInputQubits}: The input object.

quantum_state: The quantum state containing the qubit.

qubit_index: The index of the qubit in the quantum state.

qubit_input_value::Float64: The input value for the qubit, which determines the phase.

Examples

qubit = ComputationQubit()
input = InputQubits()
quantum_state = createQureg(1, env)
qubit_index = 1
qubit_input_value = 0.5
initialise_qubit(qubit, input, quantum_state, qubit_index, qubit_input_value)

RobustBlindVerification.rand_k_0_7 — Method

Draw random interger between 0 and 7

RobustBlindVerification.separate_each_color — Method

Extracts from a Graphs.Coloring{Int64} -> Vector{Vector{Int64}} Once a coloring is selected a vector of integers will result 1 -> Dummy vertex 2 -> Trap

RobustBlindVerification.set_io_qubits_type! — Method

In Computation round, there are sometimes input values for qubits,
when this happens, this function will allocate space for them in the property graph

RobustBlindVerification.set_io_qubits_type! — Method

For MBQC, with no blind

RobustBlindVerification.set_io_qubits_type! — Method

In Test rounds there is no classical input, but this holder function alows for 
unilateral call, regardless of round.

RobustBlindVerification.set_vertex_type! — Method

Must be implemented after the round type is implemented

RobustBlindVerification.update_measurement — Method

update_measurement(client::Client, q, mg, outcome)

This function retrieves the round type from the measurement graph and then calls the update_measurement function with the client, round type, qubit, measurement graph, and outcome as arguments.

Arguments

client::Client: The client for which the measurement is being updated.

q: The qubit for which the measurement is being updated.

mg: The measurement graph associated with the computation round.

outcome: The outcome of the measurement.

Returns

The result of calling update_measurement with the client, round type, qubit, measurement graph, and outcome.

Examples

client = Client()
q = 1
mg = MeasurementGraph()
outcome = 0
update_measurement(client, q, mg, outcome)

RobustBlindVerification.update_measurement — Method

update_measurement(client::Client, round::ComputationRound, q, mg, outcome)

Update the measurement for a given computation round and qubit. The function retrieves the one-time pad integer associated with the qubit and computes the absolute difference between the outcome and the one-time pad integer.

Arguments

client::Client: The client for which the measurement is being updated.

round::ComputationRound: The computation round for which the measurement is being updated.

q: The qubit for which the measurement is being updated.

mg: The measurement graph associated with the computation round.

outcome: The outcome of the measurement.

Returns

The absolute difference between the outcome and the one-time pad integer.

Examples

client = Client()
round = ComputationRound()
q = 1
mg = MeasurementGraph()
outcome = 0
update_measurement(client, round, q, mg, outcome)

RobustBlindVerification.update_measurement — Method

update_measurement(client::Client, round::Union{MBQC, TestRound}, q, mg, outcome)

This function is used in the context of Measurement-Based Quantum Computing (MBQC) or during a test round. It takes an outcome and simply returns it without making any modifications.

Arguments

client::Client: The client for which the measurement is being updated.

round::Union{MBQC, TestRound}: Specifies whether the computation round is a part of MBQC or a test round.

q: The qubit for which the measurement is being updated.

mg: The measurement graph associated with the computation round.

outcome: The outcome of the measurement.

Returns

The same outcome that was passed as an argument.

Examples

client = Client()
round = MBQC()
q = 1
mg = MeasurementGraph()
outcome = 0
update_measurement(client, round, q, mg, outcome)

RobustBlindVerification.ClusterState — Type

ClusterState

A struct representing the cluster state in the MBQC framework.

# Description
`ClusterState` is a marker struct used to represent the cluster state in an MBQC computation. It can be used in combination with other data structures or algorithms specific to the cluster state model.

# Example
```julia
# Declare cluster state
cluster_state = ClusterState()
```

RobustBlindVerification.InputQubits — Type

InputQubits

A struct representing input qubits in the MBQC framework.

# Description
`InputQubits` is a marker struct used to indicate the presence of input qubits in an MBQC computation. It is typically used in combination with other data structures or algorithms to handle input qubits in the computation.

# Example
```julia
# Declare input qubits
input_qubits = InputQubits()
```

RobustBlindVerification.MBQCAngles — Type

MBQCAngles(angles) 

- Struct representing the angles associated to the graph. The number of angles will be the same as the vertices in the graph.

# Parameters
- `angles`: A listed set of values, as long as the angles are indexable in the same ways that the vertices are.

# Example
```
julia> angles = [π,π/4,5π/4,7π/4]
julia> mbqc_angles = MBQCAngles(angles) 
```

RobustBlindVerification.MBQCFlow — Type

MBQCFlow(forward_flow, backward_flow)

Struct representing flow in MBQC.

# Definition of Flow
- `forward_flow`, `f`: Oᶜ → Iᶜ is a mapping `v ↦ f(v)` with an inverse `f⁻¹(v) ↦ v`, with partial order "≤". The partial order is said to map the present to the future or the present to the past.
- (a) `v ∼ f(v)`, where "∼" defines the neighbourhood `N(f(v))` and `v` has set membership.
- (b) `v ≤ f(v)`
- (c) `w ∼ f(v)`, then ∀ `v`, `v ≤ w`

## Example
- One dimensional lattice, the "path graph", where a vertex is represented as "()" and an edge is represented as "---".
- Let `i` be the index of each vertex so that `i = {1, 2, 3, 4}` and
- `p := (1)---(2)---(3)---(4)`
- `f(i) := i + 1`
- `f⁻¹(i) := i - 1`
- `f([1, 2, 3]) = [2, 3, 4]`, since 4 has no future, there is no 4 + 1 answer.
- `f⁻¹([2, 3, 4]) = [1, 2, 3]`, since 1 has no past, there is no 1 - 1 answer.

# Parameters
- `forward_flow`: A mapping to take an input vertex and return the output vertex such that the definitions of flow hold. The forward flow function can be any container that takes a vertex index as input and outputs a new vertex index.
- `backward_flow`: A mapping to take an output vertex and return the input vertex. The backward flow function maps the inverse of the forward flow.

# Example
```julia
# Define forward and backward flow functions
forward_flow(io) = io[2]
backward_flow(io) = io[1]

# Create an MBQCFlow
mbqc_flow = MBQCFlow(forward_flow, backward_flow)
```

RobustBlindVerification.MBQCGraph — Type

MBQCGraph(graph,input,output)

- Struct representing the graph used in the MBQC. Container holds the graph as well as the input and output sets.

# Parameters
- `graph`: Any graph suitable for MBQC
- `input`: has type MBQCInput
- `output`: had type MBQCOutput

# Example
```
julia> using Graphs # use using Pkg; Pkg.add("Graphs") is not installede
julia> graph = Graphs.grid([1,4]) # 1D cluster graph (path graph) on 4 vertices
julia> indices,values = (1),(0)
julia> input  = MBQCInput(indices,values)
julia> indices = (4)
julia> output  = MBQCOutput(indices)
julia> mbqc_graph = MBQCGraph(graph,input,output)
```

RobustBlindVerification.MBQCInput — Type

MBQCInput(indices,values)

- Struct representing an input set into the graph, can be empty

# Parameters
- `indices`: has type Tuple on normally integers (whole numbers 1 to N) and correspond to vertices in a graph.
- `values`: has type Tuple, can be any type

# Example
```
julia> indices = (1,2,3,4)
julia> values = (0,1,1,0) #Computational basis outcomes
julia> mbqc_input = MBQCInput(indices,values)
```

RobustBlindVerification.MBQCMeasurementOutcomes — Type

MBQCMeasurementOutcomes(outcomes)

Struct representing measurement outcomes in MBQC.

# Parameter
- `outcomes`: An array or container representing the measurement outcomes.

## Example
```julia
# Define measurement outcomes
outcomes = [0, 1, 1, 0]

# Create an MBQCMeasurementOutcomes
measurement_outcomes = MBQCMeasurementOutcomes(outcomes)
```

RobustBlindVerification.MBQCOutput — Type

MBQCOutput(indices)

- Struct representing an output set into the graph, can be empty.

# Parameters
- `indices`: has type Tuple on normally integers (whole numbers 1 to N) and correspond to the vertices in a graph.

# Example
```
julia> indices = (10,11,12)
julia> mbqc_output = MBQCOutput(indices)
```

RobustBlindVerification.MBQCResourceState — Type

MBQCResourceState(graph, flow, angles)

Struct representing a resource state in MBQC.

# Parameters
- `graph`: An instance of `MBQCGraph` representing the underlying graph structure.
- `flow`: An instance of `MBQCFlow` representing the flow in the resource state.
- `angles`: An instance of `MBQCAngles` representing the angles associated with each vertex.

## Example
```julia
# Create an MBQCGraph
graph = MBQCGraph([1, 2, 3], [(1, 2), (2, 3)])

# Create an MBQCFlow
flow = MBQCFlow((1, 2) => 2, (2, 3) => 3)

# Create an MBQCAngles
angles = MBQCAngles([π/2, π/4, π/3])

# Create an MBQCResourceState
resource_state = MBQCResourceState(graph, flow, angles)
```

RobustBlindVerification.NoInputQubits — Type

NoInputQubits

A struct representing the absence of input qubits in the MBQC framework.

# Description
`NoInputQubits` is a marker struct used to indicate the absence of input qubits in an MBQC computation. It can be used as a flag or placeholder to handle scenarios where there are no input qubits in the computation.

# Example
```julia
# Declare absence of input qubits
no_input_qubits = NoInputQubits()
```

RobustBlindVerification.NoisyClient — Type

    NoisyClient(noise_model::Union{Vector{NoiseModel}, NoiseModel})

For MBQC only. A structure representing a client that operates under some noise model. May be buggy.

Fields

noise_model::Union{Vector{NoiseModel}, NoiseModel}: The noise model under which the client operates. This can be a single NoiseModel or a vector of NoiseModels.

Examples

noise_model = NoiseModel()
client = NoisyClient(noise_model)