Robust Blind Verification

Package documentation

Docs for package

RobustBlindVerification.MetaGraph — Method

MetaGraph(::Client, resource::MBQCResourceState)

This function creates a MetaGraph from a given MBQCResourceState. It extracts the graph from the resource state and wraps it in a MetaGraph.

Arguments

client::Client: The client object.

resource::MBQCResourceState: The MBQC resource state containing the graph.

Examples

julia client = Client() resource = MBQCResourceState(graph) MetaGraph(client, resource)`

RobustBlindVerification.add_correction_vertices! — Method

add_correction_vertices!(::Client, mg, resource::MBQCResourceState)

This function adds correction vertices to the meta graph for a client in the MBQC model. It iterates over each vertex in the resource, gets the correction vertices for each, and sets these as properties in the meta graph.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the properties will be added.

resource::MBQCResourceState: The resource containing the graph and its coloring.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
add_correction_vertices!(client, mg, resource)

RobustBlindVerification.add_flow_vertex! — Method

add_flow_vertex!(::Client, mg, resource::MBQCResourceState, flow_type::Union{ForwardFlow,BackwardFlow})

This function adds a flow vertex to the meta graph for a client in the MBQC model. It first converts the flow type to a symbol, then iterates over each vertex in the resource. For each vertex, it gets the verified flow output and sets this as a property in the meta graph.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the properties will be added.

resource::MBQCResourceState: The resource containing the graph and its coloring.

flow_type::Union{ForwardFlow,BackwardFlow}: The flow type to be added.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
flow_type = ForwardFlow()
add_flow_vertex!(client, mg, resource, flow_type)

RobustBlindVerification.add_flow_vertex! — Method

add_flow_vertex!(::Client, mg, resource::MBQCResourceState)

This function adds both forward and backward flow vertices to the meta graph for a client in the MBQC model. It calls the add_flow_vertex! function twice, once with ForwardFlow and once with BackwardFlow.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the properties will be added.

resource::MBQCResourceState: The resource containing the graph and its coloring.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
add_flow_vertex!(client, mg, resource)

RobustBlindVerification.add_noise! — Method

add_noise!(client::Client, noise_model::NoiseModel)

This function adds noise to a quantum system according to the specified noise model. It calls the add_noise! function with a new client and the given noise model as arguments.

Arguments

client::Client: The client for which the noise is being added.

noise_model::NoiseModel: The noise model to be used.

Examples

client = Client()
noise_model = NoiseModel()
add_noise!(client, noise_model)

RobustBlindVerification.add_noise! — Method

add_noise!(client::Client, model::Union{Damping,Dephasing,Depolarising,Pauli}, params::QubitNoiseParameters)

This function adds noise to a quantum system. The noise model can be one of Damping, Dephasing, Depolarising, or Pauli. The function throws an error if the noise type is not SingleQubit. It then iterates over the range of qubits represented in the backend and adds noise to each qubit according to the specified model and parameters.

Arguments

client::Client: The client for which the noise is being added.

model::Union{Damping,Dephasing,Depolarising,Pauli}: The noise model to be used.

params::QubitNoiseParameters: The parameters for the noise model.

Examples

client = Client()
model = Damping()
params = QubitNoiseParameters(SingleQubit(), backend)
add_noise!(client, model, params)

RobustBlindVerification.add_output_qubits! — Method

add_output_qubits!(::Client, mg, resource::MBQCResourceState)

This function adds output qubits to the meta graph for a client in the MBQC model. It retrieves the output indices from the resource graph and sets the :output_inds property of the meta graph to these indices.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the property will be added.

resource::MBQCResourceState: The resource containing the graph and its coloring.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
add_output_qubits!(client, mg, resource)

RobustBlindVerification.add_round_type! — Method

add_round_type!(::Client, mg, round_type)

This function adds a round type to the meta graph for a client in the MBQC model. It sets the :round_type property of the meta graph to the specified round type.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the property will be added.

round_type: The round type to be added to the meta graph.

Returns

The updated MetaGraph.

Examples

client = Client()
mg = MetaGraphs.MetaGraph(graph)
round_type = "round1"
add_round_type!(client, mg, round_type)

RobustBlindVerification.c_iterator — Method

c_iterator(N)

Create a C based index iterator that generates numbers from 0 to N-1.

Arguments

N: The upper limit for the circular iterator.

Returns

A circular iterator that generates numbers from 0 to N-1.

Examples

# Create a circular iterator
N = 5
iterator = c_iterator(N)

RobustBlindVerification.c_shift_index — Method

c_shift_index(n::Int)

Compute the shifted index n-1 for circular indexing.

Arguments

n::Int: The input index.

Returns

The shifted index n-1 for circular indexing.

Examples

# Compute the shifted index
n = 3
shifted_index = c_shift_index(n)

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Computation ∩ Qubit ∈ Input set → δᵥ = ϕᵥ + (θᵥ + xᵥπ) + rᵥπ

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Computation ∩ Qubit ∉ Input set → δᵥ = ϕᵥ′ + θᵥ + rᵥπ

RobustBlindVerification.compute_angle_δᵥ — Method

MBQC compute updated angle

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Test ∩ Qubit ≡ Dummy → δᵥ = {kπ/r | k ∼ U(0..7)}

RobustBlindVerification.compute_angle_δᵥ — Method

Computation of δᵥ Case

Round ≡ Test ∩ Qubit ≡ Trap → δᵥ = θᵥ + rᵥπ

RobustBlindVerification.compute_backward_flow — Method

compute_backward_flow(graph, forward_flow, vertex)

This function computes the backward flow of a given vertex in a graph. It first finds the neighbors of the vertex and checks if the vertex is in the forward flow of any of its neighbors. If it is not, the function returns 0. If it is, the function finds the index of the vertex in the forward flow of its neighbors. If the vertex is not in the flow of the neighbors, an error is thrown. If there is more than one past vertex found, an error is thrown. Otherwise, the function returns the first past vertex.

Arguments

graph: The graph.

forward_flow: The forward flow function.

vertex: The vertex for which to compute the backward flow.

Returns

The first past vertex if it exists, 0 otherwise.

Examples

graph = Graph(5)
forward_flow = (n -> n + 1)
vertex = 3
backward_flow_vertex = compute_backward_flow(graph, forward_flow, vertex)

RobustBlindVerification.convert_flow_type_symbol — Method

convert_flow_type_symbol(::Client, flow_type::Union{ForwardFlow,BackwardFlow})

This function converts a flow type (either ForwardFlow or BackwardFlow) into a symbol. The conversion process involves converting the flow type to a string, removing parentheses, adding an underscore before "Flow", converting to lowercase, and finally converting to a symbol.

Arguments

::Client: The Client object.

flow_type::Union{ForwardFlow,BackwardFlow}: The flow type to be converted.

Returns

A Symbol representing the flow type.

Examples

client = Client()
flow_type = ForwardFlow()
flow_sym = convert_flow_type_symbol(client, flow_type)

RobustBlindVerification.create_quantum_env — Method

create_quantum_env(client::Client)

Create a new quantum environment using the QuEST environment creation function.

Arguments

client::Client: The client for which the quantum environment is being created.

Returns

A new quantum environment.

Examples

client = Client()
env = create_quantum_env(client)

RobustBlindVerification.create_quantum_state — Method

create_quantum_state(client::Client, density_matrix::DensityMatrix, quantum_env, num_qubits)

Create a quantum state using a density matrix representation.

Arguments

client::Client: The client object representing the quantum computing service.

density_matrix::DensityMatrix: The density matrix object representing the quantum state.

quantum_env: The quantum environment object.

num_qubits: The number of qubits to be used in the quantum state.

Returns

The quantum state represented as a density matrix.

Examples

client = Client()
env = create_quantum_env(client)
state = create_quantum_state(client, DensityMatrix(), env, 2)

RobustBlindVerification.create_quantum_state — Method

create_quantum_state(client::Client, state::StateVector, quantum_env, num_qubits)

Create a new quantum state as a state vector using the QuEST function createQureg.

Arguments

client::Client: The client for which the quantum state is being created.

state::StateVector: Indicates that the quantum state should be created as a state vector.

quantum_env: The quantum environment in which the quantum state is being created.

num_qubits: The number of qubits in the quantum state.

Returns

A new quantum state as a state vector.

Examples

client = Client()
env = create_quantum_env(client)
state = create_quantum_state(client, StateVector(), env, 2)

RobustBlindVerification.draw_bit — Method

Draw random bit 0 or 1

RobustBlindVerification.draw_dᵥ — Method

Draw random bit 0 or 1 for dummy

RobustBlindVerification.draw_rᵥ — Method

Draw random bit 0 or 1 for trap

RobustBlindVerification.draw_θᵥ — Method

For draw θ a multiple of kπ/4, k ∈ 0..7

RobustBlindVerification.entangle_graph! — Method

entangle_graph!(::Client, mg)

This function entangles the quantum state of a meta graph for a client in the MBQC model. It first retrieves the quantum state from the meta graph and creates a graph from the meta graph. Then, for each edge in the graph, it applies a controlled phase flip operation on the source and destination vertices.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the properties will be added.

Examples

client = Client()
mg = MetaGraphs.MetaGraph(graph)
entangle_graph!(client, mg)

RobustBlindVerification.generate_property_graph! — Method

generate_property_graph!(::Client, round_type, resource::MBQCResourceState, state_type::Union{StateVector,DensityMatrix})

This function generates a property graph for a client in the MBQC model based on the round type. It first creates a meta graph from the resource and adds the round type to it. Then, it adds output qubits, sets the vertex type and IO qubits type, initializes the qubits, adds flow vertices and correction vertices, initializes measurement outcomes, and initializes the quantum state of the meta graph. This function is run at the beginning of every round.

Arguments

::Client: The Client object.

round_type: The round type.

resource::MBQCResourceState: The resource containing the graph and its coloring.

state_type::Union{StateVector,DensityMatrix}: The type of quantum state to create.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
round_type = "round1"
state_type = StateVector()
generate_property_graph!(client, round_type, resource, state_type)

RobustBlindVerification.generate_property_graph! — Method

generate_property_graph!(::Client, round_type::MBQC, resource::MBQCResourceState, state_type::Union{StateVector,DensityMatrix})

This function generates a property graph for a client in the MBQC model based on the round type. It first creates a meta graph from the resource and adds the round type to it. Then, it adds output qubits, sets the vertex type and IO qubits type, initializes the qubits, adds flow vertices and correction vertices, initializes measurement outcomes, initializes the quantum state of the meta graph, and entangles the graph. This function is run at the beginning of every round.

Arguments

::Client: The Client object.

round_type::MBQC: The round type.

resource::MBQCResourceState: The resource containing the graph and its coloring.

state_type::Union{StateVector,DensityMatrix}: The type of quantum state to create.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
round_type = MBQC()
state_type = StateVector()
generate_property_graph!(client, round_type, resource, state_type)

RobustBlindVerification.generate_random_greedy_color — Method

generate_random_greedy_color(g, reps)

This function generates a random greedy coloring of a graph. It uses the random_greedy_color function from the Graphs package.

Arguments

g: The graph to be colored.

reps: The number of repetitions for the random greedy coloring algorithm.

Examples

g = Graphs.grid_graph((5,5))
reps = 10
generate_random_greedy_color(g, reps)

RobustBlindVerification.get_edge_iterator — Method

get_edge_iterator(resource::MBQCResourceState)

Retrieve an iterator over the edges in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.

# Returns
An iterator over the edges of the resource state's graph.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the edge iterator
edge_iterator = get_edge_iterator(resource)
```

RobustBlindVerification.get_number_vertices — Method

get_number_qubits(resource::MBQCResourceState)

Retrieve the number of qubits in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.

# Returns
The number of qubits in the resource state.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the number of qubits
num_qubits = get_number_qubits(resource)
```

RobustBlindVerification.get_random_coloring — Method

get_random_coloring(c::Vector{Vector{Int64}})

This function selects a random coloring from a vector of colorings.

Arguments

c::Vector{Vector{Int64}}: The vector of colorings.

Examples

julia c = [[1, 2, 1], [2, 1, 2], [1, 2, 2]] get_random_coloring(c)`

RobustBlindVerification.get_vertex_iterator — Method

get_vertex_iterator(resource::MBQCResourceState)

Retrieve an iterator over the vertices in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.

# Returns
An iterator over the vertices of the resource state's graph.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the vertex iterator
vertex_iterator = get_vertex_iterator(resource)
```

RobustBlindVerification.get_vertex_neighbours — Method

get_vertex_neighbours(resource::MBQCResourceState, vertex)

Retrieve the neighbors of a given vertex in an MBQC resource state.

# Arguments
- `resource::MBQCResourceState`: An MBQC resource state containing the graph representation of the resource.
- `vertex`: The vertex for which to retrieve the neighbors.

# Returns
An array of vertices representing the neighbors of the specified vertex.

# Examples
```julia
# Create an MBQC resource state
graph = MBQCGraph(...)
flow = MBQCFlow(...)
angles = MBQCAngles(...)
resource = MBQCResourceState(graph, flow, angles)

# Get the neighbors of a vertex
vertex = 1
neighbors = get_vertex_neighbours(resource, vertex)
```

RobustBlindVerification.init_measurement_outcomes! — Method

init_measurement_outcomes!(::Client, mg, resource::MBQCResourceState)

This function initializes the measurement outcomes in the meta graph for a client in the MBQC model. It iterates over each vertex in the resource and sets the :outcome property of each vertex in the meta graph to Int64.

Arguments

::Client: The Client object.

mg: The MetaGraph to which the properties will be added.

resource::MBQCResourceState: The resource containing the graph and its coloring.

Returns

The updated MetaGraph.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
init_measurement_outcomes!(client, mg, resource)

RobustBlindVerification.init_plus_phase_state! — Method

init_plus_phase_state!(nophase::NoPhase, qureg, qᵢ)

This function initializes a quantum state to a superposition state (|+⟩ state) without adding any phase. It applies a Hadamard gate to the qubit, putting it into a superposition state.

Arguments

nophase::NoPhase: The NoPhase object indicating that no phase is to be added.

qureg: The quantum register containing the qubit.

qᵢ: The index of the qubit in the quantum register.

Examples

nophase = NoPhase()
qureg = createQureg(1, env)
qᵢ = 1
init_plus_phase_state!(nophase, qureg, qᵢ)

RobustBlindVerification.init_plus_phase_state! — Method

init_plus_phase_state!(phase::Phase, qureg, qᵢ, φᵢ)

This function initializes a quantum state to a superposition state (|+⟩ state) with a specified phase. It first applies a Hadamard gate to the qubit, putting it into a superposition state. Then it applies a Z rotation to the qubit, adding the specified phase.

Arguments

phase::Phase: The phase object.

qureg: The quantum register containing the qubit.

qᵢ: The index of the qubit in the quantum register.

φᵢ: The phase to be added to the qubit.

Examples

phase = Phase()
qureg = createQureg(1, env)
qᵢ = 1
φᵢ = π/2
init_plus_phase_state!(phase, qureg, qᵢ, φᵢ)

RobustBlindVerification.init_qubit — Method

init_qubit(::DummyQubit)::Int64

This function is used to initialize the qubit in the meta graph. The state is not given, but the bit for the initial state of the dummy qubit is returned.

Arguments

dummy::DummyQubit: The DummyQubit object.

Returns

An Int64 representing the bit for the initial state of the dummy qubit.

Examples

julia dummy = DummyQubit() bit = init_qubit(dummy)`

RobustBlindVerification.init_qubit — Method

init_qubit(::TrapQubit)::Float64

This function is used to initialize the qubit in the meta graph. The state is not given, but the angle for the plus phase state for a trap qubit is returned.

Arguments

trap::TrapQubit: The TrapQubit object.

Returns

A Float64 representing the angle for the plus phase state for a trap qubit.

Examples

trap = TrapQubit()
angle = init_qubit(trap)

RobustBlindVerification.init_qubit_meta_graph! — Method

init_qubit_meta_graph!(::Client, resource, mg)

This function initializes the qubit meta graph for a client in the MBQC model. It retrieves the round type from the meta graph and then calls the appropriate init_qubit_meta_graph! function based on the round type.

Arguments

::Client: The Client object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the properties will be added.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_prop!(mg, :round_type, ComputationRound())
init_qubit_meta_graph!(client, resource, mg)

RobustBlindVerification.init_qubit_meta_graph! — Method

init_qubit_meta_graph!(::Client, ::ComputationRound, resource::MBQCResourceState, mg)

This function initializes the qubit meta graph for a client in the MBQC model during a computation round. It sets the secret angle and initial qubit properties for each vertex in the meta graph.

Arguments

::Client: The Client object.

::ComputationRound: The ComputationRound object.

resource::MBQCResourceState: The resource containing the graph and its coloring.

mg: The MetaGraph to which the properties will be added.

Examples

client = Client()
round = ComputationRound()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
init_qubit_meta_graph!(client, round, resource, mg)

RobustBlindVerification.init_qubit_meta_graph! — Method

init_qubit_meta_graph!(::Client, mbqc::MBQC, resource::MBQCResourceState, mg)

This function initializes the qubit meta graph for a client in the MBQC model. It sets the secret angle and initial qubit properties for each vertex in the meta graph.

Arguments

client::Client: The Client object.

mbqc::MBQC: The MBQC object.

resource::MBQCResourceState: The resource containing the graph and its coloring.

mg: The MetaGraph to which the properties will be added.

Examples

client = Client()
mbqc = MBQC()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
init_qubit_meta_graph!(client, mbqc, resource, mg)

RobustBlindVerification.init_qubit_meta_graph! — Method

init_qubit_meta_graph!(::Client, ::TestRound, resource::MBQCResourceState, mg)

This function initializes the qubit meta graph for a client in the MBQC model during a test round. It sets the initial qubit property for each vertex in the meta graph based on the vertex type.

Arguments

::Client: The Client object.

::TestRound: The TestRound object.

resource::MBQCResourceState: The resource containing the graph and its coloring.

mg: The MetaGraph to which the properties will be added.

Examples

client = Client()
round = TestRound()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
init_qubit_meta_graph!(client, round, resource, mg)

RobustBlindVerification.initialise_quantum_state_meta_graph! — Method

initialise_quantum_state_meta_graph!(::Client, state_type::Union{StateVector,DensityMatrix}, mg)

This function initializes the quantum state of a meta graph for a client in the MBQC model. It first creates a quantum environment and a quantum state of the specified type. Then, for each vertex in the meta graph, it gets the vertex type, vertex IO type, and initial qubit value, and uses these to initialize the qubit in the quantum state. Finally, it sets the quantum state as a property of the meta graph.

Arguments

::Client: The Client object.

state_type::Union{StateVector,DensityMatrix}: The type of quantum state to create.

mg: The MetaGraph to which the properties will be added.

Returns

The updated MetaGraph.

Examples

client = Client()
mg = MetaGraphs.MetaGraph(graph)
state_type = StateVector()
initialise_quantum_state_meta_graph!(client, state_type, mg)

RobustBlindVerification.initialise_quantum_state_meta_graph! — Method

initialise_quantum_state_meta_graph!(mbqc::MBQC, client::Client, state_type::Union{StateVector,DensityMatrix}, mg)

This function initializes the quantum state of a meta graph for a client in the MBQC model. It first creates a quantum environment and a quantum state of the specified type. Then, for each vertex in the meta graph, it gets the vertex type and vertex IO type, and uses these to initialize the qubit in the quantum state. Finally, it sets the quantum state as a property of the meta graph.

Arguments

::MBQC: The MBQC object.

::Client: The Client object.

state_type::Union{StateVector,DensityMatrix}: The type of quantum state to create.

mg: The MetaGraph to which the properties will be added.

Returns

The updated MetaGraph.

Examples

mbqc = MBQC()
client = Client()
mg = MetaGraphs.MetaGraph(graph)
state_type = StateVector()
initialise_quantum_state_meta_graph!(mbqc, client, state_type, mg)

RobustBlindVerification.initialise_qubit — Method

initialise_qubit(dummy::DummyQubit, noinput::NoInputQubits, quantum_state, qubit_index, qubit_input_value::Int)

This function initialises a qubit in a quantum state. If the input value for the qubit is zero, it does nothing. If the input value is one, it applies a Pauli-X gate to the qubit. If the input value is neither zero nor one, it throws a DummyQubitZeroOneInitialisationError.

Arguments

dummy::DummyQubit: The DummyQubit object.

noinput::NoInputQubits: The NoInputQubits object.

quantum_state: The quantum state containing the qubit.

qubit_index: The index of the qubit in the quantum state.

qubit_input_value::Int: The input value for the qubit.

Examples

dummy = DummyQubit()
noinput = NoInputQubits()
quantum_state = createQureg(1, env)
qubit_index = 1
qubit_input_value = 0
initialise_qubit(dummy, noinput, quantum_state, qubit_index, qubit_input_value)

RobustBlindVerification.initialise_qubit — Method

initialise_qubit(mbqc::MBQC, qubit::Union{ComputationQubit,TrapQubit}, input::Union{InputQubits,InputQubits,NoInputQubits}, quantum_state, qubit_index)

This function initialises a qubit in a quantum state to a superposition state without adding any phase. It applies a Hadamard gate to the qubit, putting it into a superposition state.

Arguments

mbqc::MBQC: The MBQC object.

qubit::Union{ComputationQubit,TrapQubit}: The type of the qubit.

input::Union{InputQubits,InputQubits,NoInputQubits}: The input object.

quantum_state: The quantum state containing the qubit.

qubit_index: The index of the qubit in the quantum state.

Examples

mbqc = MBQC()
qubit = ComputationQubit()
input = InputQubits()
quantum_state = createQureg(1, env)
qubit_index = 1
initialise_qubit(mbqc, qubit, input, quantum_state, qubit_index)

RobustBlindVerification.initialise_qubit — Method

initialise_qubit(qubit::Union{ComputationQubit,TrapQubit}, input::Union{InputQubits,InputQubits,NoInputQubits}, quantum_state, qubit_index, qubit_input_value::Float64)

This function initialises a qubit in a quantum state with a phase determined by the input value. If the input value is a float, it initialises the qubit to a superposition state with the input value as the phase. If the input value is not a float, it throws a QubitFloatPhaseInitialisationError.

Arguments

qubit::Union{ComputationQubit,TrapQubit}: The type of the qubit.

input::Union{InputQubits,InputQubits,NoInputQubits}: The input object.

quantum_state: The quantum state containing the qubit.

qubit_index: The index of the qubit in the quantum state.

qubit_input_value::Float64: The input value for the qubit, which determines the phase.

Examples

qubit = ComputationQubit()
input = InputQubits()
quantum_state = createQureg(1, env)
qubit_index = 1
qubit_input_value = 0.5
initialise_qubit(qubit, input, quantum_state, qubit_index, qubit_input_value)

RobustBlindVerification.measure_along_ϕ_basis! — Method

measure_along_ϕ_basis!(client::Client, ψ, v::Union{Int32,Int64}, ϕ::Float64)

This function measures a quantum state along a specific basis. It first applies a Z rotation to the state, then applies a Hadamard gate, and finally performs a measurement. The basis is determined by the angle ϕ.

Arguments

client::Client: The Client object.

ψ: The quantum state to be measured.

v::Union{Int32,Int64}: The vertex on which the operations are applied.

ϕ::Float64: The angle determining the basis for measurement.

Returns

The result of the measurement.

Examples

client = Client()
ψ = QuantumState()
v = 1
ϕ = π/4
measure_along_ϕ_basis!(client, ψ, v, ϕ)

RobustBlindVerification.produce_initialised_graph — Method

produce_initialised_graph(::Client, mg)

This function produces an initialised graph from a meta graph for a client in the MBQC model. It simply creates a new graph from the meta graph.

Arguments

::Client: The Client object.

mg: The MetaGraph to be converted into a graph.

Returns

A new Graph created from the MetaGraph.

Examples

client = Client()
mg = MetaGraphs.MetaGraph(graph)
produce_initialised_graph(client, mg)

RobustBlindVerification.produce_initialised_qureg — Method

produce_initialised_qureg(client::Client, mg)

This function retrieves the quantum state from a meta graph for a client in the MBQC model. It uses the get_prop function to get the :quantum_state property from the meta graph.

Arguments

client::Client: The Client object.

mg: The MetaGraph from which the quantum state will be retrieved.

Returns

The quantum state of the MetaGraph.

Examples

client = Client()
mg = MetaGraphs.MetaGraph(graph)
produce_initialised_qureg(client, mg)

RobustBlindVerification.rand_k_0_7 — Method

Draw random interger between 0 and 7

RobustBlindVerification.separate_each_color — Method

separate_each_color(g::Graphs.Coloring{Int64})

This function extracts from a Graphs.Coloring{Int64} and returns a Vector{Vector{Int64}}. Once a coloring is selected, a vector of integers will result where:

1 represents a Dummy vertex

2 represents a Trap

Arguments

g::Graphs.Coloring{Int64}: The graph coloring object.

Examples

julia g = Graphs.grid_graph((5,5)) coloring = Graphs.greedy_color(g) separate_each_color(coloring)`

RobustBlindVerification.set_io_qubits_type! — Method

set_io_qubits_type!(::Client, resource, mg)

This function sets the input/output qubits type property in a MetaGraph based on the round type property of the MetaGraph.

Arguments

client::Client: The Client object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex io type property will be added.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_prop!(mg, :round_type, ComputationRound())
set_io_qubits_type!(client, resource, mg)

RobustBlindVerification.set_io_qubits_type! — Method

set_io_qubits_type!(::ComputationRound, resource, mg)

In Computation round, there are sometimes input values for qubits. When this happens, this function will allocate space for them in the property graph. It assigns the InputQubits type to the vertices that are in the input indices and NoInputQubits to the rest.

Arguments

round::ComputationRound: The ComputationRound object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex io type property will be added.

Examples

round = ComputationRound()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_io_qubits_type!(round, resource, mg)

RobustBlindVerification.set_io_qubits_type! — Method

set_io_qubits_type!(::MBQC, resource, mg)

This function sets the input/output qubits type property in a MetaGraph for MBQC with no blind. It assigns the InputQubits type to the vertices that are in the input indices and NoInputQubits to the rest.

Arguments

client::MBQC: The MBQC client object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex io type property will be added.

Examples

client = MBQC()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_io_qubits_type!(client, resource, mg)

RobustBlindVerification.set_io_qubits_type! — Method

set_io_qubits_type!(::TestRound, resource, mg)

In Test rounds there is no classical input, but this holder function allows for unilateral call, regardless of round. It assigns the NoInputQubits type to all the vertices.

Arguments

round::TestRound: The TestRound object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex io type property will be added.

Examples

round = TestRound()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_io_qubits_type!(round, resource, mg)

RobustBlindVerification.set_vertex_type! — Method

set_vertex_type!(::Client, resource, mg)

This function sets the vertex type property in a MetaGraph based on the round type property of the MetaGraph. It must be implemented after the round type is implemented.

Arguments

client::Client: The Client object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex type property will be added.

Examples

client = Client()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_prop!(mg, :round_type, ComputationRound())
set_vertex_type!(client, resource, mg)

RobustBlindVerification.set_vertex_type! — Method

set_vertex_type!(::TestRound, resource, mg)

This function sets the vertex type property in a MetaGraph based on a random color pattern from the test round in the resource graph. It assigns the DummyQubit and TrapQubit types to the vertices according to the color pattern.

Arguments

test_round::TestRound: The TestRound object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex type property will be added.

Examples

test_round = TestRound()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_vertex_type!(test_round, resource, mg)

RobustBlindVerification.set_vertex_type! — Method

set_vertex_type!(::Union{MBQC,ComputationRound}, resource, mg)

This function sets the vertex type property in a MetaGraph based on the color pattern of the computation round in the resource graph. It assigns the ComputationQubit type to the vertices according to the color pattern.

Arguments

mbqc::Union{MBQC,ComputationRound}: The MBQC or ComputationRound object.

resource: The resource containing the graph and its coloring.

mg: The MetaGraph to which the vertex type property will be added.

Examples

mbqc = MBQC()
resource = MBQCResourceState(graph)
mg = MetaGraphs.MetaGraph(resource.graph.graph)
set_vertex_type!(mbqc, resource, mg)

RobustBlindVerification.store_measurement_outcome! — Method

store_measurement_outcome!(client::Client, client_meta_graph, qubit, outcome)

This function stores the measurement outcome of a specific qubit in the client's meta graph. It uses the set_p function to set the property in the meta graph.

Arguments

client::Client: The Client object.

client_meta_graph: The MetaGraph where the measurement outcome will be stored.

qubit: The qubit whose measurement outcome is being stored.

outcome: The measurement outcome to be stored.

Returns

Nothing. The function modifies the clientmetagraph in-place.

Examples

client = Client()
client_meta_graph = MetaGraphs.MetaGraph(graph)
qubit = 1
outcome = 0
store_measurement_outcome!(client, client_meta_graph, qubit, outcome)

RobustBlindVerification.update_measurement — Method

update_measurement(client::Client, q, mg, outcome)

This function retrieves the round type from the measurement graph and then calls the update_measurement function with the client, round type, qubit, measurement graph, and outcome as arguments.

Arguments

client::Client: The client for which the measurement is being updated.

q: The qubit for which the measurement is being updated.

mg: The measurement graph associated with the computation round.

outcome: The outcome of the measurement.

Returns

The result of calling update_measurement with the client, round type, qubit, measurement graph, and outcome.

Examples

client = Client()
q = 1
mg = MeasurementGraph()
outcome = 0
update_measurement(client, q, mg, outcome)

RobustBlindVerification.update_measurement — Method

update_measurement(client::Client, round::ComputationRound, q, mg, outcome)

Update the measurement for a given computation round and qubit. The function retrieves the one-time pad integer associated with the qubit and computes the absolute difference between the outcome and the one-time pad integer.

Arguments

client::Client: The client for which the measurement is being updated.

round::ComputationRound: The computation round for which the measurement is being updated.

q: The qubit for which the measurement is being updated.

mg: The measurement graph associated with the computation round.

outcome: The outcome of the measurement.

Returns

The absolute difference between the outcome and the one-time pad integer.

Examples

client = Client()
round = ComputationRound()
q = 1
mg = MeasurementGraph()
outcome = 0
update_measurement(client, round, q, mg, outcome)

RobustBlindVerification.update_measurement — Method

update_measurement(client::Client, round::Union{MBQC, TestRound}, q, mg, outcome)

This function is used in the context of Measurement-Based Quantum Computing (MBQC) or during a test round. It takes an outcome and simply returns it without making any modifications.

Arguments

client::Client: The client for which the measurement is being updated.

round::Union{MBQC, TestRound}: Specifies whether the computation round is a part of MBQC or a test round.

q: The qubit for which the measurement is being updated.

mg: The measurement graph associated with the computation round.

outcome: The outcome of the measurement.

Returns

The same outcome that was passed as an argument.

Examples

client = Client()
round = MBQC()
q = 1
mg = MeasurementGraph()
outcome = 0
update_measurement(client, round, q, mg, outcome)

RobustBlindVerification.ClusterState — Type

ClusterState

A struct representing the cluster state in the MBQC framework.

# Description
`ClusterState` is a marker struct used to represent the cluster state in an MBQC computation. It can be used in combination with other data structures or algorithms specific to the cluster state model.

# Example
```julia
# Declare cluster state
cluster_state = ClusterState()
```

RobustBlindVerification.InputQubits — Type

InputQubits

A struct representing input qubits in the MBQC framework.

# Description
`InputQubits` is a marker struct used to indicate the presence of input qubits in an MBQC computation. It is typically used in combination with other data structures or algorithms to handle input qubits in the computation.

# Example
```julia
# Declare input qubits
input_qubits = InputQubits()
```

RobustBlindVerification.MBQCAngles — Type

MBQCAngles(angles) 

- Struct representing the angles associated to the graph. The number of angles will be the same as the vertices in the graph.

# Parameters
- `angles`: A listed set of values, as long as the angles are indexable in the same ways that the vertices are.

# Example
```
julia> angles = [π,π/4,5π/4,7π/4]
julia> mbqc_angles = MBQCAngles(angles) 
```

RobustBlindVerification.MBQCFlow — Type

MBQCFlow(forward_flow, backward_flow)

Struct representing flow in MBQC.

# Definition of Flow
- `forward_flow`, `f`: Oᶜ → Iᶜ is a mapping `v ↦ f(v)` with an inverse `f⁻¹(v) ↦ v`, with partial order "≤". The partial order is said to map the present to the future or the present to the past.
- (a) `v ∼ f(v)`, where "∼" defines the neighbourhood `N(f(v))` and `v` has set membership.
- (b) `v ≤ f(v)`
- (c) `w ∼ f(v)`, then ∀ `v`, `v ≤ w`

## Example
- One dimensional lattice, the "path graph", where a vertex is represented as "()" and an edge is represented as "---".
- Let `i` be the index of each vertex so that `i = {1, 2, 3, 4}` and
- `p := (1)---(2)---(3)---(4)`
- `f(i) := i + 1`
- `f⁻¹(i) := i - 1`
- `f([1, 2, 3]) = [2, 3, 4]`, since 4 has no future, there is no 4 + 1 answer.
- `f⁻¹([2, 3, 4]) = [1, 2, 3]`, since 1 has no past, there is no 1 - 1 answer.

# Parameters
- `forward_flow`: A mapping to take an input vertex and return the output vertex such that the definitions of flow hold. The forward flow function can be any container that takes a vertex index as input and outputs a new vertex index.
- `backward_flow`: A mapping to take an output vertex and return the input vertex. The backward flow function maps the inverse of the forward flow.

# Example
```julia
# Define forward and backward flow functions
forward_flow(io) = io[2]
backward_flow(io) = io[1]

# Create an MBQCFlow
mbqc_flow = MBQCFlow(forward_flow, backward_flow)
```

RobustBlindVerification.MBQCGraph — Type

MBQCGraph(graph,input,output)

- Struct representing the graph used in the MBQC. Container holds the graph as well as the input and output sets.

# Parameters
- `graph`: Any graph suitable for MBQC
- `input`: has type MBQCInput
- `output`: had type MBQCOutput

# Example
```
julia> using Graphs # use using Pkg; Pkg.add("Graphs") is not installede
julia> graph = Graphs.grid([1,4]) # 1D cluster graph (path graph) on 4 vertices
julia> indices,values = (1),(0)
julia> input  = MBQCInput(indices,values)
julia> indices = (4)
julia> output  = MBQCOutput(indices)
julia> mbqc_graph = MBQCGraph(graph,input,output)
```

RobustBlindVerification.MBQCInput — Type

MBQCInput(indices,values)

- Struct representing an input set into the graph, can be empty

# Parameters
- `indices`: has type Tuple on normally integers (whole numbers 1 to N) and correspond to vertices in a graph.
- `values`: has type Tuple, can be any type

# Example
```
julia> indices = (1,2,3,4)
julia> values = (0,1,1,0) #Computational basis outcomes
julia> mbqc_input = MBQCInput(indices,values)
```

RobustBlindVerification.MBQCMeasurementOutcomes — Type

MBQCMeasurementOutcomes(outcomes)

Struct representing measurement outcomes in MBQC.

# Parameter
- `outcomes`: An array or container representing the measurement outcomes.

## Example
```julia
# Define measurement outcomes
outcomes = [0, 1, 1, 0]

# Create an MBQCMeasurementOutcomes
measurement_outcomes = MBQCMeasurementOutcomes(outcomes)
```

RobustBlindVerification.MBQCOutput — Type

MBQCOutput(indices)

- Struct representing an output set into the graph, can be empty.

# Parameters
- `indices`: has type Tuple on normally integers (whole numbers 1 to N) and correspond to the vertices in a graph.

# Example
```
julia> indices = (10,11,12)
julia> mbqc_output = MBQCOutput(indices)
```

RobustBlindVerification.MBQCResourceState — Type

MBQCResourceState(graph, flow, angles)

Struct representing a resource state in MBQC.

# Parameters
- `graph`: An instance of `MBQCGraph` representing the underlying graph structure.
- `flow`: An instance of `MBQCFlow` representing the flow in the resource state.
- `angles`: An instance of `MBQCAngles` representing the angles associated with each vertex.

## Example
```julia
# Create an MBQCGraph
graph = MBQCGraph([1, 2, 3], [(1, 2), (2, 3)])

# Create an MBQCFlow
flow = MBQCFlow((1, 2) => 2, (2, 3) => 3)

# Create an MBQCAngles
angles = MBQCAngles([π/2, π/4, π/3])

# Create an MBQCResourceState
resource_state = MBQCResourceState(graph, flow, angles)
```

RobustBlindVerification.NoInputQubits — Type

NoInputQubits

A struct representing the absence of input qubits in the MBQC framework.

# Description
`NoInputQubits` is a marker struct used to indicate the absence of input qubits in an MBQC computation. It can be used as a flag or placeholder to handle scenarios where there are no input qubits in the computation.

# Example
```julia
# Declare absence of input qubits
no_input_qubits = NoInputQubits()
```

RobustBlindVerification.NoisyClient — Type

    NoisyClient(noise_model::Union{Vector{NoiseModel}, NoiseModel})

For MBQC only. A structure representing a client that operates under some noise model. May be buggy.

Fields

noise_model::Union{Vector{NoiseModel}, NoiseModel}: The noise model under which the client operates. This can be a single NoiseModel or a vector of NoiseModels.

Examples

noise_model = NoiseModel()
client = NoisyClient(noise_model)